renamed some components and files
